Rag_Project/requirements.txt
google-generativeai
chromadb
pdfplumber
gradio
